Add a slerp (spherical interpolation) move for orbits etc.
This is still pretty hacky, but it works.  The idea is that you can slerp every aspect of the view state, and you can also center it about some point so that you can rotate about that point.

This is not meant to be necessarily immediately user-usable, but rather a more foundational move for e.g. orbiting around a point or rotating a camera.  You can use e.g. Rotations.AngleAxis to set the final point and then this will go there but on the surface of a sphere parameterized by the original and final points, and the center point.